examples/libsmbclient: Cast mode_t to unsigned int for GNU/Solaris build

Autobuild-User(master): Jeremy Allison <jra@samba.org>
Autobuild-Date(master): Fri Mar 15 20:22:08 CET 2013 on sn-devel-104

+++ examples/libsmbclient/testchmod.c
@@ -42,7 +42,7 @@ int main(int argc, char * argv[])
         return 1;
     }
     
-    printf("\nBefore chmod: mode = %04o\n", st.st_mode);
+    printf("\nBefore chmod: mode = %04o\n", (unsigned int)st.st_mode);
     
     if (smbc_chmod(pSmbPath, mode) < 0)
     {
@@ -56,7 +56,7 @@ int main(int argc, char * argv[])
         return 1;
     }
     
-    printf("After chmod: mode = %04o\n", st.st_mode);
+    printf("After chmod: mode = %04o\n", (unsigned int)st.st_mode);
     
     return 0; 
 }
